requirements.txt
azure-ai-ml
azure-identity
azureml-core
mlflow
pandas
pillow
scikit-learn